+++ CHANGELOG.md
@@ -1,7 +1,21 @@
 # Changelog
 
+---
+
 Todos los cambios notables en este proyecto serán documentados en este archivo
 de acuerdo a los estándares de [Keep a Changelog](https://keepachangelog.com/es-ES/).
+
+---
+
+## [v0.1.0] - 2026-05-23
+
+### Added
+
+- Implementación inicial de lógica del proyecto
+- Consulta a la API de Steam para descarga de datos
+- Creacion de interfaz grafica de usuario con CustomTkinter
+
+---
 
 ## [v0.0.1] - 2026-05-22